added users + Update backend/models/post.js and backend/routes/commentsRoute.js to include username field in comments model

+++ backend/routes/postRoute.js
@@ -1,6 +1,8 @@
 const express = require('express');
 const router = express.Router();
 const { Post } = require('../models');
+const { authenticationHandler } = require('../middlewares/authenticationHandler');
+
 
 router.get("/", async (req, res) => {
   try {
@@ -11,20 +13,16 @@ router.get("/", async (req, res) => {
   }
 });
 
-router.post("/", async (req, res) => {
-  try {
-    const post = req.body;
-    const newPost = await Post.create(post);
-    res.json(newPost);
-  } catch (error) {
-    res.status(500).json({ error: error.message });
-  }
-});
-
-router.put("/:id", async (req, res) => {
+router.put("/:id", authenticationHandler, async (req, res) => {
   try {
     const postId = req.params.id;
     const updatedPost = req.body;
+    const post = await Post.findOne({ where: { id: postId } });
+
+    if (post.UserId !== req.user.id) {
+      return res.status(403).json({ error: 'You are not authorized to edit this post' });
+    }
+
     await Post.update(updatedPost, { where: { id: postId } });
     res.json(updatedPost);
   } catch (error) {
@@ -32,9 +30,15 @@ router.put("/:id", async (req, res) => {
   }
 });
 
-router.delete("/:id", async (req, res) => {
+router.delete("/:id", authenticationHandler, async (req, res) => {
   try {
     const postId = req.params.id;
+    const post = await Post.findOne({ where: { id: postId } });
+
+    if (post.UserId !== req.user.id) {
+      return res.status(403).json({ error: 'You are not authorized to delete this post' });
+    }
+
     await Post.destroy({ where: { id: postId } });
     res.json({ message: 'Post deleted successfully' });
   } catch (error) {
